feat: Update execute function to handle larger integers

+++ src/lib.rs
@@ -3,7 +3,7 @@
 //! Bu kod, Bilişsel İşletim Sistemi'nden tamamen bağımsız olarak derlenir.
 
 #[no_mangle]
-pub extern "C" fn execute(input: i32) -> i32 {
-    // Ajanın eylemi: Gelen sayının karesini alır (Örn: 6 -> 36)
+pub extern "C" fn execute(input: i64) -> i64 {
+    // Artık devasa sayılar çökmeyecek
     input * input
 }
